Only send ammatillisen kielikoe results to sure if there are changes
Otherwise the "myonnetty" date for existing results will get updated
when any scores are saved from valinta-ui.
This still leaves the problem that if the value is changed, the new
arvosana gets the current date, which might not be what is wanted
when correcting past results.
Then again, this is probably not really fixable unless we add the
"real" date for the exam somewhere (e.g. the parameters of the haku).