Lazy Observable handling without replay
It is better to use Observables in a more idiomatic way,
explicitly subscribing them to make things happen, and
use replay explicitly only when needed.

As a lot of existing code relies on one or both of these
features, we should not change it directly; hence the
duplicated methods along the old ones.

If this approach seems to work, we can try to change
single places of use to use the new methods little by
little, when touching those parts of code for other
reasons.

I will try to employ this in fixing BUG-1610 .

No functional changes.