BUG-1610 : Single case CAS HTTP 302 retry
Observable.retry methods do not do anything useful
when wrapping current "eager", replaying versions of
the HttpResourceImpl created Observables, because
the original error will simply be replayed on retries.
Determining whether the exception is caused by a CAS
redirect in the HttpResourceImpl level is not trivial,
so this information is kludged from CasKoosteInterceptor
to HttpResourceImpl with a custom HTTP header added
to the inbound message.

Furthermore, as Observable.retry methods consume the
exceptions causing the original failures, they must be
thrown explicitly when not retrying. It would be great
to find a cleaner way to do it. But testing this
behaviour is tricky because you should always consider
all these different cases:
* first request succeeds
* first request fails on CAS redirect, some subsequent
  request succeeds
* first request and retries fail on CAS redirect
* first request fails on other problem => no retry

In the failure cases, we want to get the exception
stack trace on the application log, and an error to
the caller.

Let's try things out with this and then start changing
other parts accordingly if this works out (or fix as
needed).